fix pid

+++ exp/pid/src/pid.c
@@ -7,12 +7,12 @@ void pid_init(float DELTA_T, float KP, float KI, float KD) {
     kd = KD;
 }
 
-float pid_calc(float result, float target, float error[2], double *integral) {
+float pid_calc(int16_t result, int16_t target, float error[2], float *integral) {
     float p, i, d;
 
     error[0] = error[1];
     error[1] = result - target;
-    *integral += (((error[0] + error[1]) / 2.0) * deltat);
+    *integral += (((float)(error[0] + error[1]) / 2.0) * deltat);
 
     p = kp * error[1];
     i = ki * (*integral);
